TODO: don't retry tasks at the workqueue level. let parsl retrying handle things. maybe put in a configurable option to let WQ retries happen that defaults to not.
Some retries already don't get retried by WQ (as evidenced by the test suite)
- I think that parsl level task failures are regarded as "successful" by WQ itself
(so that the exception is successfully sent up to the parsl level?)

in which case wq level retries happen eg in the case of a node failure.

that makes it much less interesting to ever set this to anything other than 0,
I think? (maybe it interacts with the auto classification stuff?)