experiment: workflow-pluggable retry scoring
when a task fails, instead of decrementing retry count by 1 (repeatedly down to 0)
instead allow a workflow-pluggable piece of code to inspect the error (with some
context) and decide how much to decrease the retry count.

other design space points:

 - code could decide to retry or not, and the retry count is a scratch record it gets
   but parsl doesn't touch that retry count at all.

 - the retry scorer, if given the task record, is not prohibited from fiddling with
   the task record. What sort of things are plausible there?
    - changing the executor label for the task, so that the job retries somewhere
      else - for example some executor that is more expensive (slower to schedule
      but longer tasks?)
    - retry for a time period rather than for a number of counts - eg so as to not
         burn through retry units very fast in a fast failing transient case -
         so "quick fails" are discounted

There's now a distinction between fail count (as in the number of times we've tried
to run this task), and something like "fail cost" - the amount of badness that this
task has accumulated (instead of accumulating 1 unit each time). These two quantities
should probably be tracked independently (including in monitoring)

As we're plugging in user code, exceptions in this logic can be expected to break
(as in, user code is "not trusted") so we need to handle retry logic exceptions
rather than just hanging the DFK.
TEST CASE: a retry handler that always raises an exception. (this should result
in the app invocation failing)

TEST CASE: a retry handler that always returns 1 (rather than the default,
which behaves the same)

TEST CASE: a retry handler that returns 0.5 and check that we get more retries

TEST CASE: a retry handler that retries for a minute