Allow all required resources to be ommitted in WQ resource spec.
The requirement is either that all of them are supplied, or none of
them are supplied.

Prior to use of parsl_resource_specification for additional specs,
the "or none of them" part of this could be achieved by omitting
the resource specification entirely. That doesn't work now that
other resource specs can be added.