Make parsl wait for ThreadPoolExecutor tasks to complete during shutdown
When parsl shuts down at the end of a python process, previously it would
not wait for thread pool tasks to complete.

However, the underlying thread pool executor (from the concurrent.* library)
would then cause the python process exit to stall until thread-local tasks
are complete.

This PR makes parsl also wait, so that it is clearer when debugging slow
or hanging shutdown that thread-local tasks are involved.

In the case of shutting down a DFK at process exit, this should not
significantly delay shutdown any futher.

In the case of shutting down a DFK in the middle of a process, this will
cause longer delay, but in the absence of the ability to cancel tasks
in the thread local executor, this maybe the best thing to do (rather than
claiming that shutdown has actually completed when tasks are still
running)