avoid 50msg/sec rate limit in monitoring router for DFK messages
DFK can send many more messages than that, and they queue up in the router
process making it use a lot of memory.

quentin experienced this. i am able to reproduce with a tight loop submitting
200000 tasks.

before this patch:
there is a delay of loop freq (* some multiple) for every DFK message received,
if other non-dfk receives have to timeout.

with this patch:
the DFK receive loops tight until there are no more messages to receive. on my
laptop this means the router uses hardly any extra memory (750mb compared to
550mb at router startup)

other receives should be audited for doing this loop-to-dequeue-without-block
behaviour: at least the interchange receiver will show the same behaviour.

however, perhaps the interchange and DFK receiving pieces should be merged
together: if messages are properly tagged and using the same syntax, then they
don't need to be separate, and that's a change I'm working on earlier in
the stack