new pip behaviour in CI is making things fail with this error and refusal to install
This is a practical problem if wanting to have both doc option and
montioring option installed in the same place: Jinja 3.0 errors out
when trying to build the docs.

ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
2379flask 2.0.0 requires Jinja2>=3.0, but you have jinja2 2.11.3 which is incompatible.
2380Successfully installed jinja2-2.11.3 parsl-1.1.0+desc.2021.5.14b
2381The command "pip install .[docs]" exited with 0.

hack on that...