Regularise NODE_INFO message format
Prior to this, the monitoring router reformatted NODE_INFO messages.
This PR regularises the format of NODE_INFO messages at the point
that they are initially sent, so that the router does not have to
reformat them.

The router still adds a run_id field to the NODE_INFO message, as
before, as this information is not available in the interchange where
the messages are generated

This moves around these non-routing behaviours in MonitoringRouter:

* reformatting an htex last_heartbeat time into a more suitable type
* putting the message timestamp into the message dictionary