Workflow-pluggable retry scoring (#2068)
When a task fails, instead of causing a retry "cost" of 1 (the previous behaviour), this PR allows that cost to be determined by a user specified function which is given some context about the failure.

After this PR there is a distinction between the accumulated retry cost, and the number of tries (which were previously the same number). This change is reflected in the monitoring database, with both values available.