Fix CI broken by dependency package changes (#2105)
* tickle CI

* Remove `coverage` test requirement dependency

This is only needed by pytest-cov, which brings in the
dependency anyway.

The version pin on `coverage` that is removed was conflicting with
pytest-cov > 2.10. Previously pip was able to easily find a solution
to that by using an earlier pytest-cov, but as of some recent
ecosystem change, it was unable to find a solution.

When manually trying to install pytest-cov 2.11.0 on my laptop
with a working parsl dependency stack already installed:
====
The conflict is caused by:
    The user requested coverage==4.5.4
    pytest-cov 2.11.0 depends on coverage>=5.2.1
====

When building in CI, this resulted in pip spending "forever" (i.e. longer
than the test timeout) trying to find a solution.

* Add .coveragerc to see if this fixes possible coverage concurrency problem

* remove blank line