Pull os x multiprocessing code into a single module (#2099)
There are a few pieces of os x compatibility code in the parsl codebase which deal with multiprocessing.{Process, Queue}.

This PR attempts to consolidate them in one place.

Most parsl processes need to be launching using multiprocessing fork method, no matter which platform.

This PR introduces ForkProcess, which will always fork, no matter which platform (or fail, when forking is not available).
Because this always forks, there is no need to switch behaviour based on operating system type.

This PR also introduces SizedQueue which is a Queue which can have its size taken. The default linux queue implements this directly, but a wrapper is needed for os x. Using SizedQueue will work on both platforms.

This should also make parsl more resistant to the global multiprocessing process type being set to anything other than fork (which happens by default on os x but is also possible on other platforms).

Co-authored-by: yongyanrao <yongyan.rao@gmail.com>
Co-authored-by: benc <benc@8740.local>
Co-authored-by: Yadu Nand Babuji <yadudoc1729@gmail.com>