Abstract more block handling from HighThroughputExecutor and share with WorkQueue (#2071)
Several (most) executors past and present (ipp, htex, and in this PR, WorkQueue) make use of "blocks": a collection of worker processes launched using Providers, Launchers and Channels usually into some kind of batch system - but not all executors, though.

The StatusHandlingExecutor provides a base class that executor implementations to give some common error handling behaviour related to blocks.

This PR moves more block handling code from the high throughput executor in that base class, and renames it BlockProviderExecutor. The immediate intention is to allow the block handling code from the high throughput executor to be re-used by the WorkQueue executor.

This pushes subclasses of the (former) StatusHandlingExecutor to use blocks in the same way as the high throughput executor does. I think that is fine. An executor can still be implemented without any parsl support for block-style scaling (as happens with the ThreadPoolExecutor) and can still make use of providers, launchers and channels if desired.

After making that abstraction, this PR then uses BlockProviderExecutor to allow the workqueue executor to use the block system in a similar way to HighThroughputExecutor. This has been tested with the DESC project for a couple of months.

As with some uses of HighThroughputExecutor, but even more so due to per-task resource specification, when using WorkQueue the scaling code doesn't always know the number of simultaneous tasks that can be executed in a block.