Re-enabled scale down test
This was disabled due to frequent failure (see issue #1885) but this test
has been re-enabled in the `desc` branch for a while and has not caused
trouble, so this commit should re-enabled it in master.

Issue #1885 also notes the utility of this test in discovering scaling-in
breakage - which gives some pressure to re-enable this test.