Fix htex scale down breakage due to overly aggressive result heartbeat
PR #2104 introduced a result channel heartbeat. However, this heartbeat
resulted in the idle time of managers being updated on each heartbeat.

Because of this, no manager ever appears to be idling for very long.

And because of this, the scale down code will not select any manager
for scaling down.

This PR makes result channel messages not update the idle time if
a manager already has an idle time: a result channel message (either
a result or a heartbeat) now means:
"I am (still) idle", rather than "I am idle starting from now".