getting this error
    head = f.read(JobStatus.SUMMARY_TRUNCATION_THRESHOLD / 2)
TypeError: argument should be integer or None, not 'float'

in an lsst run

That means the status file is >1024 bytes to go down that
code path.

What is that file in case of the local provider?

looks liek there's a python2 -> python3 change here:

python2   1/2 gives 0, <type 'int'>A
>>> type(1/2)
<type 'int'>


python3   1/2 gives 0.5
>>> type(1/2)
<class 'float'>