Make monitoring hub exit condition more explicit (#2131)
Prior to this PR, the monitoring hub decided it was time to exit
by relying on the absence of a python version string in WORKFLOW_INFO
messages that it was sent.

This PR replaces that with an explicit exit field, which makes
the shutdown logic easier to read. It should not change any other
behaviour.