Tidy bash error codes tests. (#2130)
The bash formatting test is removed, as that implicit formatting was
removed in PR #1034

Re-enable div_0 bash error code test. This was disabled as part of the
large set of tests disabled in the PR #652 testpocalypse. But on review
it should be a valid test that we expect to pass.

Co-authored-by: Yadu Nand Babuji <yadudoc1729@gmail.com>