catch invalid state exceptions when setting whole-executer failures
these happen if the set all invalid function is called more than once.
it's unclear if the intention is that function is only called once,
or if this function should be impervious to being called more than once.

it might cause a hang problem if a second future is added to the executor
*after* it has been called once (so that it is erroring after the first
future, and never setting the second future at all).

this might be quite a niche race condition but maybe reproducible?

this changes in this patch at the moment turn this into an ERROR with a
stack trace - that's maybe not the right reporting if tihs is expected.