Make changes for CI reliability (#2118)
This CI makes a couple of changes for CI reliability:

Recently CI tests have been taking longer than 40 minutes of walltime, and this PR extends the 40 minute time limit to 60 minutes to give them more chance to succeed (unless they're taking that long because of a hang)

This PR also copies runinfo/ from test runs. This is for debugging ongoing hangs in --config local tests that I have been unable to reproduce on my laptop, but which occurred in travis CI before and are still occuring in github actions.