Allocate block ids uniquely (#2182)
Previously the code around block ID allocation was a bit fuzzy and
it looks like it could sometimes allocate the same block ID several
times in the presence of failure.

This is at least causing confusion in debugging frequent hangs in
CI.

A second ID allocated was used when a block ID had not already been
allocated, for simulating failures. This ID allocation code is changed
in this PR to use the same sequence counter.