CI fixing megablast (#2205)
CI was down to less than 50% of builds passing, and then to 100% failure with a new release of pandas.

This PR contains a number of distinct changes which I have tried for fixing and debugging CI, none on which were in themselves completely reliable.

This PR has passed 6 times in a row with no code changes, so the reliability looks like it is massively increased. I still think there is some broken behaviour, as the reason that 6 is only 6 is because the 7th to last test hung, and all that's changed since then is addition of some logging around the suspicious area.

* Make retry tests wait for all launched tasks before finishing

Without this, the tasks continue to run while subsequent local
tests happen. This is at the least confusing in logs - I haven't
checked if it introduces any actual testing problems.

* More aggressively shut down DFKs in  tests

* Format BadStateException with the executor label rather than full repr

The full repr for an executor is usually very long and includes all
of the configuration information for that exception, so previously
a BadStateException looked like this, which is awkwardly large:

parsl.executors.errors.BadStateException: Executor HighThroughputExecutor(
    address=None,
    address_probe_timeout=None,
    cores_per_worker=1,
    cpu_affinity='none',
    heartbeat_period=30,
    heartbeat_threshold=120,
    interchange_port_range=(55000, 56000),
    label='htex_local',
    launch_cmd='executable_that_hopefully_does_not_exist_1030509.py',
    managed=True,
    max_workers=1,
    mem_per_worker=None,
    poll_period=1,
    prefetch_capacity=0,
    provider=LocalProvider(
        channel=LocalChannel(
            envs={},
            script_dir='/home/runner/work/parsl/parsl/runinfo/030/submit_scripts',
            userhome='/home/runner/work/parsl/parsl'
        ),
        cmd_timeout=30,
        init_blocks=1,
        launcher=SimpleLauncher(debug=True),
        max_blocks=1,
        min_blocks=0,
        move_files=None,
        nodes_per_block=1,
        parallelism=1,
        worker_init=''
    ),
    storage_access=None,
    worker_debug=True,
    worker_logdir_root=None,
    worker_port_range=(54000, 55000),
    worker_ports=None,
    working_dir=None
) failed due to:        STDERR: /home/runner/work/parsl/parsl/runinfo/030/submit_scripts/parsl.localprovider.1642801165.5067718.sh: line 3: executable_that_hopefully_does_not_exist_1030509.py: command not found

After this commit, the exception instead looks like this:

parsl.executors.errors.BadStateException: Executor htex_local failed due to:    STDERR: /home/benc/parsl/src/parsl/runinfo/000/submit_scripts/parsl.localprovider.1643124620.9775066.sh: line 3: executable_that_hopefully_does_not_exist_1030509.py: command not found

* Add in additional logging around shutdown, to help with ongoing misbehaving shutdown debugging

* Tidy HighThroughputExecutor shutdown function to remove incorrect commentary and unused parameters

* Explicitly shut down htex after bad state test

Without this, the interchange process continues running for the duration
of the pytest lifetime, overlapping with other tests.

This makes CI hangs harder to debug.

* turn on debug in parsl visualize to see why it is failing in CI

* Restrict pandas version used by monitoring

Pandas 1.4, just released, breaks parsl-visualize with this:

ImportError: Pandas requires version '1.4.0' or newer of 'sqlalchemy' (version '1.3.24' currently installed).

* these tests should complete in around 18 minutes... 30 mins is enough timeout for now

* turn over coverage checking as its little used and i've seen some suspicious coverage related errors

* Add debug logging around suspicious hanging behaviour in interchange in CI