run_id does not need to be configurable - there are other monitoring fields for this. remove TODO (#2212)
This value is a surrogate key generated by parsl. Reuse of runids will break lots of assumptions about runid uniqueness. It shouldn't be user configurable.

There are other fields for identifying run identity according to the users identity rules (workflow name and version), as well as the ability to use the generated runid as a key into a different user-controlled data structure.