Add type annotations for provider_base.py and fix some bugs there (#2211)
Most of these annotations have been brewing in the benc-mypy branch for a while. My recent work on exaworks psij, which has much stronger type-checking, has pushed me to want these annotations in mainstream parsl while implementing a parsl provider which interfaces with psij.

Implementing this interface with psij and using these type annotations has revealed a bug when a job status is created with a None stdout or stderr: a default None stdout or stderr path causes an exception when accessing the stdout or stderr property. This is fixed to return None in that case.