Fix test task wait behaviour that could leave tasks running
This will wait for all tasks to complete without raising exceptions,
before then raising any exceptions that need raising.

Without this, then if an exception was raised by assertion checking,
the test would not wait for other tasks to complete, and the test
suite could continue with more tests overlapping these task
completions.

This is confusing when debugging, so this PR aims to avoid that.