Put default job error handling in one place.
Prior to this commit, job error handling failures
were dealt with in two places - a "default" with 3 retries,
and for all status handling executors, in handle errors,
with an "init blocks" limit which overrode that.

This removes the default handling -- 3 is an unconfigured
arbitrary value at the moment, and in practice this code
path isn't used because our error handling executors are
all status handling executors.

Handling errors at the executor level is now mandatory,
but can still be delegated to simple_error_handler.

This is part of an ongoing move towards BlockProviderExecutor
holding the relevant code (which can then be overridden
if desired) rather than having a second non-inheritence-based
default path.