Only query active slurm jobs (#2259)
When squeue is called with at least one job ID which is active, it returns successfully. If it is called with only jobs that have already completed, it returns an error. This situation happens if no blocks are currently running.

The same logic (checking if the existing status is terminal) is used in the local and kubernetes providers.

Co-authored-by: Alec Clowes <alec.clowes@roivant.com>