Fix tasks_per_app plot
This previously did not display at all.

This PR adds in task_time_returned to the underlying SQL query, which is
always needed.

This PR also uses the workflow end time as an implied task end time, so
that tasks which do not have an explicit end time can be shown as continuing
until the end of the workflow run.