Set forked process names for debugging memory/cpu usage (#2285)
Previously, in `ps`, all processes appeared with the same commandline,
that of the process which which was used to launch parsl (for example,
python3 ..., or pytest ..., or bps ... in one LSST project).

This commit gives more interesting names.

Perhaps the process name should include some kind of workflow identifier so as to
disambiguate? That's an issue for users running multiple workflows, most likely spread
across time when leftover processes are left behind. Although the regular process
three would show relationships.

This patch deals with processes forked with multiprocessing, in:
 * monitoring
 * high throughput executor
 * work queue executor