slurm: have squeue be explicit about the format
If the user has SQUEUE_FORMAT set, the results from squeue
might not be in the expected format. Changed the code to be
explicit about this format from squeue. This also allows a
briefer format with simpler parsing because we can dump the
header and columns we don't care about.