Use a single ZMQ socket in monitoring router (#2292)
There is a significant cost to waiting for zmq messages in the way that it is
implemented in the monitoring router, because of timeouts when there is no
traffic on a particular channel.  That imposes a maximum reception rate
(eg the reciprocal of the sum of all timeouts in the receive loop) which can
cause massive queueing and memeory bloat.

Using a single ZMQ socket instead of several reduces this bloat.