Remove apparently unused vestigial checkpoint test script
It has no test_ in it so never runs in pytest, so I guess
it was originally intended to be run by hand.

Given the structure of the test it looks like it hasn't
been run by hand in a very long time - since at least
PR #335 introduced class based configuration in June 2018.