Remove awkward Actions column from viz workflow summary
This previously contained only a link to the resource summary page,
which is also linked from individual workflow summaries.

The workflow summaries page also has links to other interesting info - the
DAGs view.

It is inconsistent having only the resources report linked from the Actions
column; and this link is not an Action.

Remove Action entirely and leave everything linked from per-workflow info.