Make pytest auto-selector fail on unknown site, rather than trivially passing
The current behaviour is when site cannot be identified, to instead pick a configuration
that runs on the localhost, without being particularly clear that a site-specific
configuration is not being used.

This can lead to false test passes.

This change forces users to see that a site-specific configuration has not been
selected, rather than giving them a false test pass.