Correct units in poll_period use in worker watchdog
Previously self.poll_period which is mostly used as a number
of milliseconds, was used as a number of seconds for the
worker watchdog.

This PR changes that so that the worker watchdog runs with
the same poll period (measured in ms) as other poll_period
loops.

This will increase the CPU load of the worker watchdog,
possibly noticeably.