Change pbs qstat to use json format (#2408)
PBSPro defaults to using the Torque qstat command which is limited in the number of characters used to represent a job id. This is a problem on the ALCF Polaris machine which has long job names (id.polaris.alcf....) which are shortened in qstat (e.g., 123.polaris.al*). These shorthand job ids do not resolve in the dictionary of resources, causing the provider to fail when scaling.

This PR provides a customized PBSPro status() function that uses qstat -f -F json to retrieve the full job description as json and then extracts the long job id from the dictionary of jobs.