Restrict all formatting in conditional log message
** THIS PR CHANGES THE LOGGING BEHAVIOUR - SEE BOTTOM
OF THIS MESSAGE **

Prior to this, half the conditional formatting happened
when logging at DEBUG level; the rest of the formatting
happened always (and then was discarded by the logging
system if logging level was higher than DEBUG).

Move everything into the conditional block.

I've noticed that at peak loads, logging starts to have
a noticeable effect on throughput, and so expensive log
statements are worth dealing with.

I wonder if I can measure this in a performance test?
Somewhere where repr(arg) is expensive to compute, I expect
is where I'd notice it - which is exactly the situation that
code block is designed to fix.

***********
I'm actually unsure about this behaviour, based on the
ordering of logging.DEBUG, .INFO, etc and maybe I should
check with the original author what they intended.

before:

INFO:
  perform format,
  discard message
  (unnecessarily expensive? measurably so?)

DEBUG:
  perform format,
  log message

lower than DEBUG:
  do not perform format,
  log message
  (eg. for tracing, output all data, which might be useful)