Switch to module-based singleton for ParslSerializer
This fixes some problems where the class based serializer object was sometimes
serialized and sent over the network; there is no reasonable behaviour for
deserializing a singleton, but in the previous implementation, the caches used
for serialization were overwritten.

The ParslSerializer class is removed entirely and the parsl.serialize.facade
is the singleton instead. This is a more pythonic way of representing
singletons.

Use `git show -w` on this commit to see that most of the usual
diff is only whitespace indentation changes.

To identify places where a ParslSerializer object was deserialized, I added a
__setstate__ method which is called whenever a ParslSerializer is deserialized.

This fired when using monitoring, as monitoring carries round with it quite a
lot of serialization overhead.

@@ -171,3 +171,6 @@ class ParslSerializer(object):
         assert len(unpacked) == 3, "Unpack expects 3 buffers, got {}".format(len(unpacked))

         return unpacked
+
+    def __setstate__(self, state):
+        raise ValueError("BENC: serializer setstate called")