Refuse to serialize ParslSerialiser singleton
Deserialization of a singleton is conceptually awkward.
Prior to this patch, the state of the singleton is
overwritten by the state supplied in the serialized
singleton, which can produce surprising cache behaviour:
cached functions would "disappear" from the cache as
the registered DillSerializer is overwritten with a different
newly deserializer DillSerializer.

This patch makes serialization of ParslSerializer fail with
an exception, to make it more obvious that this undefined
behaviour is happening.

The patch that removes parsl.serializer.[de]serializer from the
imported module environment of the wrapper functions is what
makes this be no longer serialized, and after that it is
ok to apply this patch to prohibit ParslSerializer serialization.