Move per-task params out of monitoring closure for cacheability
Monitor wrapper closures are not cacheable by parsl.serialize, because they
are created on invocation of an app. parsl.serialize code caching expects that
repeated invocations of a function will use the same function code, which is
not true in this case.

The main reason that this closure is generated on each invocation is because
it contains invocation specific values: task_id and try_id.

This PR re-arranges the invocation of the monitor wrapper so that those two
values are passed as additional parameters - effectively making a manual
closure that is more amenable to parsl.serialize.

This PR introduces a new cache inside the montoring code, keyed by:

  (run_id, f, radio_mode)

which is then used to cache wrapped app functions.