A monitoring radio that sends messages at results time as part of the results package
this means that:

- in some failure cases, monitoring messages will be lost - for example, a task start message when workers die due to timeouts.

+ no need for any communications channel other than the regular results channel. in cloud-like situations (which we aren't targetting really any more?) this could be useful.

I hope even when there is a shared fs available, this will sometimes give much more efficient message delivery (with the above latency and reliability constraints). That's the main immediate driver for this work.

This reflects some of my personal opinion that a reasonable tradeoff is later reporting in exchange for more efficiency, if desired.

TODO: bug - `running` state is not being stored in the status table, even though task_time_running is being updated in the try table. Investigate what assumptions about state change are being misunderstood in the monitoring router.