More accurately record worker usage in efficiency plot
Prior to this PR, the plot was generated by computing a histogram of
one-second buckets, and each task contributed 1 second to each bucket
that it overlapped, even when the task was only running for part of
that second (the first and last seconds).

After this PR, a task only contributes a fraction of a second to the
start and end buckets, based on the fractional part of the start/end
timestamp.

For a workflow with many short running tasks, this will show a large
decrease in worker usage: for example, the pytest test suite running
with --config htex_local_alternate.py on my laptop goes from a peak
of 50 workers in use (more workers than actually exist!) to a peak
of 3.

See issue #2474 for more discussion and example plots.