Tidy resource_distribution_plot
This PR removes unused default parameters, as there are no values there that
should be favoured.

It also adds in stronger validation of those values, rather than falling
off the end of if statements or defaulting to an incorrect case.

A couple of variables named after apps are renamed after their actual use
as task identifiers - this is probably a cut and paste vestige.

Similar motivation to PR #2478