Make work queue executor pop tasks so failure handler does not see them
This is part of a fix for task #2473, but not all of it.

Prior to this PR, when Parsl decides that a work queue executor is
permanently failed, it will attempt to set an exeception on all task
futures, even tasks which already have received a result.

This results in an incorrect state exception, which then stops the
rest of executor permanent failure handling.

This change to use .pop makes the Work Queue Executor behave more
like the High Throughput Executor: completed task futures will not be
stored in the list of tasks, and so (race conditions aside) will not
get an exception set on them.